+++ web/app/package.json
@@ -66,7 +66,6 @@
     "react": "16.9.0",
     "react-avatar-editor": "11.0.7",
     "react-dom": "16.9.0",
-    "react-highlight-words": "0.16.0",
     "react-router": "5.0.1",
     "react-router-dom": "5.0.1",
     "react-transition-group": "4.2.2",
